all: return value signature changes
* return signature of all 'fops' callback changed to `gf_return_t`.

* provide IS_ERROR()/IS_SUCCESS() macros to check only this type of argument.

* provide 2 global variables `gf_success` and `gf_error` introduced, which
  can used instead of -1 and 0 return used largely today.

Updates: #280

+++ xlators/features/cloudsync/src/cloudsync-plugins/src/cvlt/src/libcvlt.h
@@ -41,7 +41,7 @@ struct _cvlt_request {
     struct iobref *iobref;
     call_frame_t *frame;
     cvlt_op_t op_type;
-    int32_t op_ret;
+    int64_t op_ret;
     int32_t op_errno;
     xlator_t *this;
     sem_t sem;
